backend/app/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
torch
scikit-learn
pandas
numpy
soundfile
python-multipart
aiofiles
websockets
python-socketio
requests
aiohttp